Test case generator
C++ codes to generate test cases in the format of the real c++ code.
Generated codes can be used directly in the program (by pasting). This
is not the best way to test the program, but is a fast and
straightforward way.

The binary file is compiled with:
Apple LLVM version 7.0.0 (clang-700.0.72)
Target: x86_64-apple-darwin15.0.0